+++ src/text/mod.rs
@@ -50,12 +50,34 @@
 //!
 //! Because the [`TextDiff::grouped_ops`] method can isolate clusters of changes
 //! this even works for very long files if paired with this method.
+//!
+//! ## Trailing Newlines
+//!
+//! When working with line diffs (and unified diffs in general) there are two
+//! "philosophies" to look at lines.  One is to diff lines without their newline
+//! character, the other is to diff with the newline character.  Typically the
+//! latter is done because text files do not _have_ to end in a newline character.
+//! As a result there is a difference between `foo\n` and `foo` as far as diffs
+//! are concerned.
+//!
+//! In similar this is handled on the [`Change`] or [`InlineChange`] level.  If
+//! a diff was created via [`TextDiff::from_lines`] the text diffing system is
+//! instructed to check if there are missing newlines encountered.  If that is
+//! the case the [`Change`] object will return true from the
+//! [`Change::missing_newline`] method so the caller knows to handle this by
+//! either rendering a virtual newline at that position or to indicate it in
+//! different ways.  For instance the unified diff code will render the special
+//! `\ No newline at end of file` marker.
 #![cfg(feature = "text")]
 use std::borrow::Cow;
 use std::cmp::Reverse;
 use std::collections::{BinaryHeap, HashMap};
+use std::fmt;
 
+mod inline;
 mod udiff;
+
+pub use self::inline::*;
 pub use self::udiff::*;
 
 use crate::algorithms::{capture_diff_slices, group_diff_ops, Algorithm, DiffOp, DiffTag};
@@ -193,8 +215,11 @@ pub struct TextDiff<'old, 'new, 'bufs> {
 /// The tag of a change.
 #[derive(Debug, PartialEq, Eq, Hash, Clone, Copy, Ord, PartialOrd)]
 pub enum ChangeTag {
+    /// The change indicates equality (not a change)
     Equal,
+    /// The change indicates deleted text.
     Delete,
+    /// The change indicates inserted text.
     Insert,
 }
 
@@ -209,6 +234,18 @@ pub struct Change<'s> {
     old_index: Option<usize>,
     new_index: Option<usize>,
     value: &'s str,
+    missing_newline: bool,
+}
+
+impl<'s> fmt::Display for Change<'s> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "{}{}",
+            self.value(),
+            if self.missing_newline { "\n" } else { "" }
+        )
+    }
 }
 
 impl<'s> Change<'s> {
@@ -232,28 +269,15 @@ impl<'s> Change<'s> {
         self.value
     }
 
-    /// Returns `true` for virtual changes.
+    /// Returns `true` if this change needs to be followed up by a
+    /// missing newline.
     ///
-    /// Virtual changes are changes that do not exist in either diff but are
-    /// necessary for a consistent user experience.  This currently only
-    /// applies to changes related to newline handling.  If lines are passed
-    /// to the [`TextDiff`] the [`TextDiff::newline_terminated`] flag is set
-    /// in which case newlines of the input are included in the changes.  However
-    /// if the trailing newline is missing it would mess up processing greatly.
-    /// Because of this a trailing virtual newline is automatically added for a
-    /// more consistent user experience.  This virtual newline can be detected
-    /// by explicitly checking for this flag.
-    pub fn is_virtual(&self) -> bool {
-        self.old_index.is_none() && self.new_index.is_none()
+    /// The [`std::fmt::Display`] implementation of [`Change`] will automatically
+    /// insert a newline after the value if this is true.
+    pub fn missing_newline(&self) -> bool {
+        self.missing_newline
     }
 }
-
-const VIRTUAL_NEWLINE_CHANGE: Change<'static> = Change {
-    tag: ChangeTag::Equal,
-    old_index: None,
-    new_index: None,
-    value: "\n",
-};
 
 impl<'old, 'new, 'bufs> TextDiff<'old, 'new, 'bufs> {
     /// Configures a text differ before diffing.
@@ -361,11 +385,6 @@ impl<'old, 'new, 'bufs> TextDiff<'old, 'new, 'bufs> {
     /// ways in which a change could be encoded (insert/delete vs replace), look
     /// up the value from the appropriate slice and also handle correct index
     /// handling.
-    ///
-    /// In addition it has some custom handling to insert "virtual" newlines
-    /// for diffs where [`TextDiff::newline_terminated`] is `true` but the
-    /// diff does not end in newlines in the right places.  For more information
-    /// see [`Change::is_virtual`].
     pub fn iter_changes(&self, op: &DiffOp) -> impl Iterator<Item = Change> {
         let newline_terminated = self.newline_terminated;
         let (tag, old_range, new_range) = op.as_tag_tuple();
@@ -373,21 +392,6 @@ impl<'old, 'new, 'bufs> TextDiff<'old, 'new, 'bufs> {
         let mut new_index = new_range.start;
         let mut old_slices = &self.old_slices()[op.old_range()];
         let mut new_slices = &self.new_slices()[op.new_range()];
-
-        // figure out if a virtual newline has to be inserted
-        let mut virtual_newline = if newline_terminated {
-            let last_element = match tag {
-                DiffTag::Equal | DiffTag::Delete | DiffTag::Replace => old_slices.last(),
-                DiffTag::Insert => new_slices.last(),
-            };
-            if !last_element.map_or(false, |x| x.ends_with(&['\r', '\n'][..])) {
-                Some(VIRTUAL_NEWLINE_CHANGE)
-            } else {
-                None
-            }
-        } else {
-            None
-        };
 
         std::iter::from_fn(move || match tag {
             DiffTag::Equal => {
@@ -400,9 +404,12 @@ impl<'old, 'new, 'bufs> TextDiff<'old, 'new, 'bufs> {
                         old_index: Some(old_index - 1),
                         new_index: Some(new_index - 1),
                         value: first,
+                        missing_newline: newline_terminated
+                            && rest.is_empty()
+                            && !first.ends_with(&['\r', '\n'][..]),
                     })
                 } else {
-                    virtual_newline.take()
+                    None
                 }
             }
             DiffTag::Delete => {
@@ -414,9 +421,12 @@ impl<'old, 'new, 'bufs> TextDiff<'old, 'new, 'bufs> {
                         old_index: Some(old_index - 1),
                         new_index: None,
                         value: first,
+                        missing_newline: newline_terminated
+                            && rest.is_empty()
+                            && !first.ends_with(&['\r', '\n'][..]),
                     })
                 } else {
-                    virtual_newline.take()
+                    None
                 }
             }
             DiffTag::Insert => {
@@ -428,9 +438,12 @@ impl<'old, 'new, 'bufs> TextDiff<'old, 'new, 'bufs> {
                         old_index: None,
                         new_index: Some(new_index - 1),
                         value: first,
+                        missing_newline: newline_terminated
+                            && rest.is_empty()
+                            && !first.ends_with(&['\r', '\n'][..]),
                     })
                 } else {
-                    virtual_newline.take()
+                    None
                 }
             }
             DiffTag::Replace => {
@@ -442,28 +455,38 @@ impl<'old, 'new, 'bufs> TextDiff<'old, 'new, 'bufs> {
                         old_index: Some(old_index - 1),
                         new_index: None,
                         value: first,
+                        missing_newline: newline_terminated
+                            && rest.is_empty()
+                            && !first.ends_with(&['\r', '\n'][..]),
                     })
-                } else if let Some(virtual_newline) = virtual_newline.take() {
-                    Some(virtual_newline)
                 } else if let Some((&first, rest)) = new_slices.split_first() {
                     new_slices = rest;
                     new_index += 1;
-                    // check for another virtual newline
-                    if newline_terminated && rest.is_empty() && !first.ends_with(&['\r', '\n'][..])
-                    {
-                        virtual_newline = Some(VIRTUAL_NEWLINE_CHANGE);
-                    }
                     Some(Change {
                         tag: ChangeTag::Insert,
                         old_index: None,
                         new_index: Some(new_index - 1),
                         value: first,
+                        missing_newline: newline_terminated
+                            && rest.is_empty()
+                            && !first.ends_with(&['\r', '\n'][..]),
                     })
                 } else {
                     None
                 }
             }
         })
+    }
+
+    /// Iterates over the changes the op expands to with inline emphasis.
+    ///
+    /// This is very similar to [`iter_changes`] but it performs a second
+    /// level per-character diff on adjacent line replacements.  The exact
+    /// behavior of this function with regards to how it detects those
+    /// inline changes is currently not defined and will likely change
+    /// over time.
+    pub fn iter_inline_changes(&self, op: &DiffOp) -> impl Iterator<Item = InlineChange> {
+        iter_inline_changes(self, op)
     }
 
     /// Returns the captured diff ops.