+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to similar are documented here.
 ## 0.5.0
 
 * Add `DiffOp::apply_to_hook` to apply a captured op to a diff hook.
-* Added virtual newline handling to `iter_changes`.
+* Added missing newline handling to the `Changes` type.
 * Made unified diff support more flexible through the introduction of
   the `UnifiedDiff` type.
 * Fixed grouped diff operation to return an empty result if the diff
   does not show any changes.
+* Added inline diff highlighting support.
 
 ## 0.4.0